Multiple compilation warnings undes MacOS fixed

+++ Tests/UnitTests/TestCore/DistributionHandlerTest.h
@@ -19,16 +19,16 @@ TEST_F(DistributionHandlerTest, DistributionHandlerConstructor)
     DistributionHandler handler;
     DistributionGate distribution(1.0, 2.0);
     std::string paraName = "value";
-    EXPECT_EQ(1, handler.getTotalNumberOfSamples());
+    EXPECT_EQ(size_t(1), handler.getTotalNumberOfSamples());
     handler.addParameterDistribution(paraName, distribution, 2, 1.0);
-    EXPECT_EQ(2, handler.getTotalNumberOfSamples());
+    EXPECT_EQ(size_t(2), handler.getTotalNumberOfSamples());
     EXPECT_EQ("DistributionHandler",handler.getName());
 
     const DistributionHandler::Distributions_t &distr = handler.getDistributions();
     ParameterDistribution distribution1 = distr[0];
     EXPECT_EQ(distribution1.getName(), "ParameterDistribution");
-    EXPECT_EQ(distribution1.getNbrSamples(),2);
-    EXPECT_EQ(distribution1.getSigmaFactor(), 1);
+    EXPECT_EQ(distribution1.getNbrSamples(), size_t(2));
+    EXPECT_EQ(distribution1.getSigmaFactor(), 1.0);
 
     ParameterPool* parameterPool = new ParameterPool();
     parameterPool->registerParameter("value",&m_value);
